+++ test/framework/src/main/java/org/elasticsearch/search/ErrorTraceHelper.java
@@ -10,22 +10,50 @@
 package org.elasticsearch.search;
 
 import org.apache.logging.log4j.Level;
+import org.elasticsearch.ExceptionsHelper;
 import org.elasticsearch.index.query.QueryShardException;
 import org.elasticsearch.test.ESIntegTestCase;
+import org.elasticsearch.test.InternalTestCluster;
 import org.elasticsearch.test.MockLog;
+import org.elasticsearch.test.transport.MockTransportService;
+import org.elasticsearch.transport.TransportMessageListener;
+import org.elasticsearch.transport.TransportService;
 
 import java.util.Arrays;
+import java.util.Optional;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.function.BooleanSupplier;
 import java.util.stream.Collectors;
 
 import static org.elasticsearch.common.Strings.format;
 import static org.elasticsearch.test.ESIntegTestCase.getNodeId;
 import static org.elasticsearch.test.ESIntegTestCase.internalCluster;
+import static org.elasticsearch.test.ESTestCase.asInstanceOf;
 
 /**
  * Utilities around testing the `error_trace` message header in search.
  */
 public enum ErrorTraceHelper {
     ;
+
+    public static BooleanSupplier setupErrorTraceListener(InternalTestCluster internalCluster) {
+        final AtomicBoolean transportMessageHasStackTrace = new AtomicBoolean(false);
+        internalCluster.getDataNodeInstances(TransportService.class)
+            .forEach(ts -> asInstanceOf(MockTransportService.class, ts).addMessageListener(new TransportMessageListener() {
+                @Override
+                public void onResponseSent(long requestId, String action, Exception error) {
+                    TransportMessageListener.super.onResponseSent(requestId, action, error);
+                    if (action.startsWith("indices:data/read/search")) {
+                        Optional<Throwable> throwable = ExceptionsHelper.unwrapCausesAndSuppressed(
+                            error,
+                            t -> t.getStackTrace().length > 0
+                        );
+                        transportMessageHasStackTrace.set(throwable.isPresent());
+                    }
+                }
+            }));
+        return transportMessageHasStackTrace::get;
+    }
 
     /**
      * Adds expectations for debug logging of a message and exception on each shard of the given index.
